requirements.txt
google-api-python-client==2.70.0
google-auth-oauthlib==0.5.3
google-auth-httplib2==0.1.0
langdetect==1.0.9
googletrans==4.0.0rc1
textblob==0.17.1
pandas==1.5.2
plotly==5.11.0
dash==2.7.0
dash-bootstrap-components==1.3.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
kaleido==0.2.1
python-dotenv==0.21.0
google-generativeai>=0.8.5

# Database and authentication dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
bcrypt==4.0.1
SQLAlchemy==1.4.46